Add option onlyAdd to Sync label command

+++ Command/Bigfoot/LabelsSyncCommand.php
@@ -28,7 +28,8 @@ class LabelsSyncCommand extends BaseCommand
             ->setName('bigfoot:labels:sync')
             ->setDefinition(array(
                 new InputArgument('target', InputArgument::OPTIONAL, 'The label dictionary file or directory', 'app/Resources/translatable_label'),
-                new InputOption('overwrite', 'o', InputOption::VALUE_NONE, 'Whether to overwrite the translations value or not (defaults to false)')
+                new InputOption('overwrite', 'o', InputOption::VALUE_NONE, 'Whether to overwrite the translations value or not (defaults to false)'),
+                new InputOption('onlyAdd', 'oA', InputOption::VALUE_NONE, 'Only add new labels (default to false)')
             ))
             ->setDescription('Synchronizes labels stored in database with those found in the file.')
             ->setHelp(<<<EOT
@@ -45,8 +46,9 @@ EOT
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $target = rtrim($input->getArgument('target'), '/');
+        $target    = rtrim($input->getArgument('target'), '/');
         $overwrite = $input->getOption('overwrite');
+        $onlyAdd   = $input->getOption('onlyAdd');
 
         if (!is_dir($target) && (!file_exists($target))) {
             throw new \InvalidArgumentException(sprintf('The target "%s" does not exist.', $input->getArgument('target')));
@@ -158,15 +160,17 @@ EOT
         }
         $em->flush();
 
-        $labels = $repo->findAll();
-        /** @var TranslatableLabel $label */
-        foreach ($labels as $label) {
-            $nameDomain = $label->getName().'-'.$label->getDomain();
-            if (!in_array($nameDomain, $processedLabels)) {
-                $em->remove($label);
+        if (!$onlyAdd) {
+            $labels = $repo->findAll();
+            /** @var TranslatableLabel $label */
+            foreach ($labels as $label) {
+                $nameDomain = $label->getName().'-'.$label->getDomain();
+                if (!in_array($nameDomain, $processedLabels)) {
+                    $em->remove($label);
+                }
             }
+            $em->flush();
         }
-        $em->flush();
 
         /** @var TranslatableLabelManager $labelManager */
         $labelManager = $this->getContainer()->get('bigfoot_core.manager.translatable_label');
